requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
httpx
python-dotenv